+++ source/RTSCamera.CommandSystem/src/Orders/VisualOrders/RTSCommandSingleVisualOrder.cs
@@ -44,7 +44,8 @@ namespace RTSCamera.CommandSystem.Orders.VisualOrders
             var selectedFormations = orderController.SelectedFormations.Where(f => f.CountOfUnitsWithoutDetachedOnes > 0).ToList();
             var orderToAdd = new OrderInQueue
             {
-                SelectedFormations = selectedFormations
+                SelectedFormations = selectedFormations,
+                ShouldAdjustFormationSpeed = Utilities.Utility.ShouldLockFormation()
             };
 
             orderToAdd.OrderType = _orderType;